move labels to recipe (???)
