+++ mealie/alembic/versions/2023-10-04-14.29.26_dded3119c1fe_added_unique_constraints.py
@@ -17,7 +17,7 @@ from sqlalchemy.orm import Session, load_only
 
 from alembic import op
 from mealie.db.models._model_utils.guid import GUID
-from mealie.db.models.labels import MultiPurposeLabel
+from mealie.db.models.recipe.labels import MultiPurposeLabel
 from mealie.db.models.recipe.ingredient import IngredientFoodModel, IngredientUnitModel, RecipeIngredientModel
 
 # revision identifiers, used by Alembic.
