+++ mealie/db/models/_all_models.py
@@ -1,5 +1,4 @@
 from .group import *
-from .labels import *
 from .recipe import *
 from .server import *
 from .users import *
