+++ mealie/db/models/recipe/ingredient.py
@@ -7,8 +7,8 @@ from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.orm.session import Session
 
 from mealie.db.models._model_base import BaseMixins, FilterableColumn, SqlAlchemyBase
-from mealie.db.models.labels import MultiPurposeLabel
 from mealie.db.models.recipe.api_extras import IngredientFoodExtras, api_extras
+from mealie.db.models.recipe.labels import MultiPurposeLabel
 
 from .._model_utils.auto_init import auto_init
 from .._model_utils.guid import GUID
