+++ mealie/db/models/recipe/labels.py
@@ -5,13 +5,13 @@ from sqlalchemy.orm import Mapped, mapped_column
 
 from mealie.db.models._model_base import BaseMixins, FilterableColumn, SqlAlchemyBase
 
-from ._model_utils.auto_init import auto_init
-from ._model_utils.guid import GUID
+from .._model_utils.auto_init import auto_init
+from .._model_utils.guid import GUID
 
 if TYPE_CHECKING:
-    from .group.group import Group
-    from .household.shopping_list import ShoppingListItem, ShoppingListMultiPurposeLabel
-    from .recipe import IngredientFoodModel
+    from ..group.group import Group
+    from ..household.shopping_list import ShoppingListItem, ShoppingListMultiPurposeLabel
+    from . import IngredientFoodModel
 
 
 class MultiPurposeLabel(SqlAlchemyBase, BaseMixins):
